add SearchHistoryRepository interface
