+++ lib/data/search_history_repository/search_history_memory_repository.dart
@@ -1,0 +1,35 @@
+import 'package:places/model/repository/search_history_repository.dart';
+
+/// Репозиторий для истории поиска
+class SearchHistoryMemoryRepository implements SearchHistoryRepository {
+  // Хранилище для списка прошлых запросов
+  final List<String> _requests = [];
+
+  /// Список запросов
+  @override
+  Future<List<String>> all() {
+    return Future.value(_requests);
+  }
+
+  /// Добавить в историю.
+  ///
+  /// Добавляется, как самый первый элемент.
+  /// Cтарые запросы стираются, чтобы не повторялись.
+  @override
+  void add(String request) {
+    remove(request);
+    _requests.insert(0, request);
+  }
+
+  /// Очистить всю историю
+  @override
+  void clear() {
+    _requests.clear();
+  }
+
+  /// Удаляет запись из истории по индексу
+  @override
+  void remove(String request) {
+    _requests.removeWhere((element) => element == request);
+  }
+}
