+++ lib/model/repository/search_history_repository.dart
@@ -1,27 +1,17 @@
 /// Репозиторий для истории поиска
-class SearchHistoryRepository {
-  // Хранилище для списка прошлых запросов
-  final List<String> _requests = [];
-
+abstract class SearchHistoryRepository {
   /// Список запросов
-  List<String> get requests => _requests;
+  Future<List<String>> all();
 
   /// Добавить в историю.
   ///
   /// Добавляется, как самый первый элемент.
   /// Cтарые запросы стираются, чтобы не повторялись.
-  void add(String request) {
-    remove(request);
-    _requests.insert(0, request);
-  }
+  void add(String request);
 
   /// Очистить всю историю
-  void clear() {
-    _requests.clear();
-  }
+  void clear();
 
   /// Удаляет запись из истории по индексу
-  void remove(String request) {
-    _requests.removeWhere((element) => element == request);
-  }
+  void remove(String request);
 }
